Prevent deletion of default user group Closes #3183

+++ tienda/admin/tables/groups.php
@@ -68,38 +68,59 @@ class TiendaTableGroups extends TiendaTable
 	 */
 	function delete($oid=null)
 	{
-		$return = parent::delete($oid);
+		$k = $this->_tbl_key;
+		$default_user_group = TiendaConfig::getInstance()->get('default_user_group', '1');
 		
-		if($return)
+		if($oid)
 		{
-			$k = $this->_tbl_key;
-			// Delete user group relationships
-			$model = JModel::getInstance('UserGroups', 'TiendaModel');
-			$model->setState('filter_group', $this->$k);
-			$links = $model->getList();
+			$key = $oid;
+		}
+		else
+		{
+			$key = $this->$k;
+		}
+		
+		
+		if( $key != $default_user_group )
+		{
+			$return = parent::delete($oid);
 			
-			if($links)
+			if($return)
 			{
-				$table = JTable::getInstance('UserGroups', 'TiendaTable');
-				foreach($links as $link)
+				
+				// Delete user group relationships
+				$model = JModel::getInstance('UserGroups', 'TiendaModel');
+				$model->setState('filter_group', $this->$k);
+				$links = $model->getList();
+				
+				if($links)
 				{
-					$table->delete($link->user_id);
+					$table = JTable::getInstance('UserGroups', 'TiendaTable');
+					foreach($links as $link)
+					{
+						$table->delete($link->user_id);
+					}
+				}
+				
+				// Delete prices
+				$model = JModel::getInstance('ProductPrices', 'TiendaModel');
+				$model->setState('filter_user_group', $this->$k);
+				$prices = $model->getList();
+				
+				if($prices)
+				{
+					$table = JTable::getInstance('ProductPrices', 'TiendaTable');
+					foreach($prices as $price)
+					{
+						$table->delete($price->user_id);
+					}
 				}
 			}
-			
-			// Delete prices
-			$model = JModel::getInstance('ProductPrices', 'TiendaModel');
-			$model->setState('filter_user_group', $this->$k);
-			$prices = $model->getList();
-			
-			if($prices)
-			{
-				$table = JTable::getInstance('ProductPrices', 'TiendaTable');
-				foreach($prices as $price)
-				{
-					$table->delete($price->user_id);
-				}
-			}
+		}
+		else
+		{
+			$this->setError(JText::_("You can't delete the default user group"));
+			return false;
 		}
 		
 		return $return;
